requirements.txt
algoliasearch-django>=4.0,<5.0
django
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
django-filter